pyproject.toml
[tool.poetry]
name = "media-organizer"
version = "0.1.0"
description = ""
authors = ["Ben Falk <falk.ben@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
textual = { git = "https://github.com/willmcgugan/textual.git", branch = "main" }
tmdbsimple = "^2.8.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"
black = "^21.7b0"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
